improved logging code setup and created socket response files
