+++ src/server_config.h
@@ -1,4 +1,4 @@
-#ifndef SERVER_CONFIG_H_
+#ifndef SERVER_CONFIG_H
 #define SERVER_CONFIG_H
 
 #define PORT 8080
@@ -7,5 +7,6 @@
 #define MAX_RESP_SIZE 150000
 #define MAX_READ_LINES 10000
 #define DEFAULT_FILE "/index.html"
+#define LOG_FILE "/var/log/http_access.log"
 
 #endif
