+++ src/socket_response.h
@@ -1,0 +1,6 @@
+#ifndef SOCKET_RESP_H
+#define SOCKET_RESP_H
+
+char *get_response_path(char *socket_response);
+
+#endif
