+++ src/utils.c
@@ -31,57 +31,3 @@ int isDirectory(const char *path) {
        return 0;
    return S_ISDIR(statbuf.st_mode);
 }
-
-char *get_response_path(char *socket_response) {
-    short int count = 0;
-    char *path = "\0";
-    char *separated_resp = NULL;
-
-    separated_resp = strtok(socket_response, " ");
-
-    while (separated_resp != NULL) {
-        if (count == 1) {
-            path = separated_resp;
-            break;
-        }
-        separated_resp = strtok(NULL, " ");
-        count++;
-    }
-
-    char *open_path = NULL;
-    open_path = malloc(
-                    sizeof(char) *
-                        (
-                            strlen(WEB_ROOT) +
-                            strlen(path) +
-                            strlen(DEFAULT_FILE) + 1
-                        )
-                    );
-
-    strcpy(open_path, WEB_ROOT);
-
-    char *absolute_path = "/";
-    if (strcmp(path, absolute_path) == 0) {
-        strcat(open_path, DEFAULT_FILE); // load root file
-    } else {
-        strcat(open_path, path);
-        if (isDirectory(open_path) != 0) {
-            strcat(open_path, DEFAULT_FILE);
-        }
-    }
-
-    return open_path;
-}
-
-
-void log_access(char *path, const char *ip, int port) {
-    FILE *logf = fopen("/var/log/http_access.log", "a");
-    if (logf == NULL) {
-        printf("Error opening file!\n");
-        exit(1);
-    }
-
-    fprintf(logf, "Requested: %s IP: %s Port: %d\r\n", path, ip, port);
-
-    fclose(logf);
-}
